refactor: AssetSettingsPage 위치 변경
