+++ src/containers/home/ScheduleDrawer/pages/AssetSettingsPage/index.tsx
@@ -1,0 +1,3 @@
+import AssetSettingsPage from "./AssetSettingsPage.tsx";
+
+export default AssetSettingsPage;
